# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/rentings": {
      "get": {
        "operationId": "getRentings",
        "summary": "",
        "parameters": [
          {
            "name": "houseTypeIds",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "array",
              "items": {
                "type": "number"
              }
            }
          },
          {
            "name": "campusIds",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "array",
              "items": {
                "type": "number"
              }
            }
          },
          {
            "name": "facilityIds",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "array",
              "items": {
                "type": "number"
              }
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "title": "PaginatedResponseOfRenting",
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginationDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Renting"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "rentings"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "RentingsController_createRenting",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRentingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "rentings"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/rentings/{rentingId}": {
      "get": {
        "operationId": "RentingsController_getRentingById",
        "parameters": [
          {
            "name": "rentingId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Renting"
                }
              }
            }
          }
        },
        "tags": [
          "rentings"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "RentingsController_updateRenting",
        "parameters": [
          {
            "name": "rentingId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRentingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "rentings"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "RentingsController_deleteRenting",
        "parameters": [
          {
            "name": "rentingId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "rentings"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/students": {
      "get": {
        "operationId": "getStudents",
        "summary": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Student"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "students"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/students/{studentId}": {
      "get": {
        "operationId": "getStudentById",
        "summary": "",
        "parameters": [
          {
            "name": "studentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Student"
                }
              }
            }
          }
        },
        "tags": [
          "students"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "updateStudent",
        "summary": "",
        "parameters": [
          {
            "name": "studentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStudentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "tags": [
          "students"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "deleteUser",
        "summary": "",
        "parameters": [
          {
            "name": "studentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "tags": [
          "students"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Veggies Api",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "RentingRes": {
        "type": "object",
        "properties": {
          "rentingId": {
            "type": "number"
          },
          "houseTypeId": {
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "landlordId": {
            "type": "number",
            "nullable": true
          },
          "campusId": {
            "type": "number",
            "nullable": true
          },
          "address": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "title": {
            "type": "string"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "square": {
            "type": "number"
          },
          "floor": {
            "type": "number"
          },
          "totalFloor": {
            "type": "number"
          },
          "isRented": {
            "type": "boolean"
          }
        },
        "required": [
          "rentingId",
          "houseTypeId",
          "createdAt",
          "landlordId",
          "campusId",
          "address",
          "price",
          "title",
          "images",
          "square",
          "floor",
          "totalFloor",
          "isRented"
        ]
      },
      "Landlord": {
        "type": "object",
        "properties": {
          "rentings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RentingRes"
            }
          },
          "landlordId": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "startTime": {
            "type": "number"
          },
          "endTime": {
            "type": "number"
          },
          "banned": {
            "type": "boolean"
          }
        },
        "required": [
          "rentings",
          "landlordId",
          "name",
          "phone",
          "email",
          "startTime",
          "endTime",
          "banned"
        ]
      },
...
      "Campus": {
        "type": "object",
        "properties": {
          "campusId": {
            "type": "number"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "campusId",
          "name"
        ]
      },
      "HouseType": {
        "type": "object",
        "properties": {
          "houseTypeId": {
            "type": "number"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "houseTypeId",
          "name"
        ]
      },
      "Facility": {
        "type": "object",
        "properties": {
          "facilityId": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "icon": {
            "type": "string"
          }
        },
        "required": [
          "facilityId",
          "name",
          "icon"
        ]
      },
...
      "PaginationDto": {
        "type": "object",
        "properties": {
          "meta": {
            "type": "object",
            "properties": {
              "page": {
                "required": true,
                "type": "number"
              },
              "limit": {
                "required": true,
                "type": "number"
              },
              "total": {
                "required": true,
                "type": "number"
              }
            }
          }
        },
        "required": [
          "meta"
        ]
      },
      "Rule": {
        "type": "object",
        "properties": {
          "ruleId": {
            "type": "number"
          },
          "content": {
            "type": "string"
          }
        },
        "required": [
          "ruleId",
          "content"
        ]
      },
      "Feature": {
        "type": "object",
        "properties": {
          "featureId": {
            "type": "number"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "featureId",
          "name"
        ]
      },
      "Renting": {
        "type": "object",
        "properties": {
          "rentingId": {
            "type": "number"
          },
          "title": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "address": {
            "type": "string"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "campus": {
            "$ref": "#/components/schemas/Campus"
          },
          "houseType": {
            "$ref": "#/components/schemas/HouseType"
          },
          "square": {
            "type": "number"
          },
          "floor": {
            "type": "number"
          },
          "totalFloor": {
            "type": "number"
          },
          "description": {
            "type": "string"
          },
          "rules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Rule"
            }
          },
          "facilities": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Facility"
            }
          },
          "features": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Feature"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "landlord": {
            "$ref": "#/components/schemas/Landlord"
          },
          "isRented": {
            "type": "boolean"
          }
        },
        "required": [
          "rentingId",
          "title",
          "price",
          "address",
          "images",
          "campus",
          "houseType",
          "square",
          "floor",
          "totalFloor",
          "description",
          "rules",
          "facilities",
          "features",
          "createdAt",
          "landlord",
          "isRented"
        ]
      },
      "CreateRentingDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 64
          },
          "description": {
            "type": "string",
            "maxLength": 128
          },
          "houseTypeId": {
            "type": "number",
            "minimum": 1
          },
          "campusId": {
            "type": "number",
            "minimum": 1
          },
          "address": {
            "type": "string",
            "maxLength": 128
          },
          "price": {
            "type": "number",
            "minimum": 1
          },
          "square": {
            "type": "number",
            "minimum": 1
          },
          "floor": {
            "type": "number",
            "minimum": 1
          },
          "totalFloor": {
            "type": "number",
            "minimum": 1
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1
            }
          },
          "features": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1
            }
          },
          "rules": {
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1
            }
          },
          "facilityIds": {
            "type": "array",
            "items": {
              "type": "number",
              "minimum": 1
            }
          }
        },
        "required": [
          "title",
          "description",
          "houseTypeId",
          "campusId",
          "address",
          "price",
          "square",
          "floor",
          "totalFloor"
        ]
      },
      "UpdateRentingDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 64
          },
          "description": {
            "type": "string",
            "maxLength": 128
          },
          "typeId": {
            "type": "number",
            "minimum": 1
          },
          "campusId": {
            "type": "number",
            "minimum": 1
          },
          "address": {
            "type": "string",
            "maxLength": 128
          },
          "price": {
            "type": "number",
            "minimum": 1
          },
          "square": {
            "type": "number",
            "minimum": 1
          },
          "floor": {
            "type": "number",
            "minimum": 1
          },
          "totalFloor": {
            "type": "number",
            "minimum": 1
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "features": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Feature"
            }
          },
          "rules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Rule"
            }
          },
          "facilityIds": {
            "type": "array",
            "items": {
              "type": "number",
              "minimum": 1
            }
          }
        }
      },
      "Favorite": {
        "type": "object",
        "properties": {
          "favoriteId": {
            "type": "number"
          },
          "studentId": {
            "type": "number"
          },
          "rentingId": {
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "favoriteId",
          "studentId",
          "rentingId",
          "createdAt"
        ]
      },
      "Student": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "studentNumber": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "favorites": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Favorite"
            }
          }
        },
        "required": [
          "studentId",
          "name",
          "studentNumber",
          "phone",
          "favorites"
        ]
      },
      "UpdateStudentDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "minLength": 6,
            "maxLength": 20
          }
        }
      },
...
    }
  }
}